Feat/#4 : Product latitude/longitude를 BigDecimal로 변경
DB 컬럼이 DECIMAL(10,7)인데 Double + @Column(precision=10, scale=7) 조합으로 매핑해서
Hibernate 6.5에서 "scale has no meaning for SQL floating point types"로 부팅이 실패하는
것을 확인함(fix/#5에서 SellerProfile에 있던 동일 패턴을 먼저 고치면서 발견, Product에도
같은 실수를 반복했었음). API DTO는 Double을 유지하고 서비스 레이어에서 변환한다.

+++ src/main/java/com/jejulocaltime/api/domain/product/Product.java
@@ -14,6 +14,7 @@ import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -95,11 +96,14 @@ public class Product {
 
     private String address;
 
+    // precision/scale은 DECIMAL(고정소수점) 컬럼에만 의미가 있어서 BigDecimal로 매핑한다.
+    // Double(부동소수점) + precision/scale 조합은 Hibernate 6.5에서
+    // "scale has no meaning for SQL floating point types"로 부팅 자체를 실패시킨다.
     @Column(precision = 10, scale = 7)
-    private Double latitude;
+    private BigDecimal latitude;
 
     @Column(precision = 10, scale = 7)
-    private Double longitude;
+    private BigDecimal longitude;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "foot_traffic_level")
